requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
djangorestframework==3.16.1
idna==3.10
mysqlclient==2.2.7
requests==2.32.5
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
python-dotenv==1.0.1
